refactor: renames, fix includes, fix meson configs
