+++ src/iguana/algorithms/Bindings.cc
@@ -1,4 +1,4 @@
-#include "AlgorithmBindings.h"
+#include "Bindings.h"
 
 #include <stdarg.h>
