+++ src/iguana/algorithms/clas12/LorentzTransformer/Algorithm.cc
@@ -1,4 +1,4 @@
-#include "LorentzTransformer.h"
+#include "Algorithm.h"
 
 // ROOT
 #include <Math/Boost.h>
