+++ src/iguana/algorithms/clas12/MomentumCorrection/Bindings.cc
@@ -1,5 +1,5 @@
-#include "MomentumCorrection.h"
-#include "iguana/algorithms/AlgorithmBindings.h"
+#include "Algorithm.h"
+#include "iguana/algorithms/Bindings.h"
 
 namespace iguana::bindings::clas12 {
   extern "C" {
