+++ src/iguana/algorithms/clas12/MomentumCorrection/Validator.cc
@@ -1,4 +1,4 @@
-#include "MomentumCorrectionValidator.h"
+#include "Validator.h"
 
 #include <TProfile.h>
 #include <TStyle.h>
