+++ src/iguana/algorithms/clas12/SectorFinder/Algorithm.cc
@@ -1,4 +1,4 @@
-#include "SectorFinder.h"
+#include "Algorithm.h"
 
 namespace iguana::clas12 {
