+++ src/iguana/algorithms/clas12/ZVertexFilter/Algorithm.cc
@@ -1,4 +1,4 @@
-#include "ZVertexFilter.h"
+#include "Algorithm.h"
 
 namespace iguana::clas12 {
