+++ src/iguana/algorithms/example/ExampleAlgorithm/Algorithm.cc
@@ -1,7 +1,7 @@
 // ############################################################################
 // # include the algorithm header
 // ############################################################################
-#include "ExampleAlgorithm.h"
+#include "Algorithm.h"
 
 // ############################################################################
 // # namespace must match that in the header
