+++ src/iguana/algorithms/physics/InclusiveKinematics/Validator.cc
@@ -1,4 +1,4 @@
-#include "InclusiveKinematicsValidator.h"
+#include "Validator.h"
 
 #include <Math/Vector3D.h>
 #include <TStyle.h>
